+++ unpythonic/syntax/__init__.py
@@ -94,6 +94,7 @@ from .lazify import *  # noqa: F401, F403
 from .letdo import *  # noqa: F401, F403
 from .letsyntax import *  # noqa: F401, F403
 from .monadic_do import *  # noqa: F401, F403
+from .multishot import *  # noqa: F401, F403
 from .nb import *  # noqa: F401, F403
 from .prefix import *  # noqa: F401, F403
 from .tailtools import *  # noqa: F401, F403